==> Urea Utilisation/code/02.GeneSetAnalysis.Rmd <==
---
title: "Gene Set Analysis"
author: "Simone Zaghen"
date: "`r format(Sys.time(), '%d %B, %Y, %H:%M')`"
output: 
  html_document: 
    toc: true
    number_section: true
    toc_float: 
      collapsed: false
      smooth_scroll: true
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../results/Reports") })

---

This script takes as input a list of comparisons (that needs to contain at least gene name, log fold change, and p value) and performs the gene set analysis using the package piano. 

# Load packages and functions
```{r, message=FALSE}
library(piano)
library(tidyverse)
library(ggpubr)

source('Functions/consGSA.R')
source('Functions/consGSAplot.R')

```

# Import data

```{r, warning=FALSE}
load("../results/Comparison_Output/Comparison_Output.RData")

```

# Gene set analysis

## Create gene set collection
Create gene set collection and subset the analysis only for different levels of the gene set collection. 

```{r}
gene_sets <- read.csv2("../data/Gene_Set.csv") #import the gene set collection

#We have three levels of analysis, and we want to perform a GSA for each
#We need to split the GSA by level
gene_sets <- list(geneSetBP = loadGSC(gene_sets %>%
                                      filter(Level %in% "biological_process") %>%
                                      select(c("genes", "Term")),
                                      type = "data.frame"),
                  geneSetMF = loadGSC(gene_sets %>%
                                      filter(Level %in% "molecular_function") %>%
                                      select(c("genes", "Term")),
                                      type = "data.frame"),
                  geneSetCC = loadGSC(gene_sets %>%
                                      filter(Level %in% "cellular_component") %>%
                                      select(c("genes", "Term")),
                                      type = "data.frame"))

```

## Perform GSA
Loop the consensus gene set analysis through all the comparisons we have. For each comparison three consensus gene set analysis will be performed (one for each level (biological process, molecular function, cellular component)). For each comparison a graph containing the GSA results for biological process, molecular function, and cellular component will be created. 

```{r, message=FALSE, warning=FALSE, results='hide'}
#create directories where to save output of GSA
if (!dir.exists("../results/Gene_Set_Analysis")){dir.create("../results/Gene_Set_Analysis")}
if (!dir.exists("../results/Gene_Set_Analysis/Plots")){dir.create("../results/Gene_Set_Analysis/Plots")}

GSA_results <- list() #Create list to store results of the consGSA function

for (i in 1:length(names(comp_list))) {
   TopTable <- as.data.frame(comp_list[i], col.names = "", row.names = 1) #create df for GSA
     GSA_results[[i]] <- list(
       BP = consGSA(TopTable, gene_sets$geneSetBP), #run consensus GSA for biological process
       MF = consGSA(TopTable, gene_sets$geneSetMF), #run consensus GSA for molecular function
       CC = consGSA(TopTable, gene_sets$geneSetCC)) #run consensus GSA for cellular component
   names(GSA_results)[i] <- names(comp_list)[i] #assign names to each element of the list
}

```

Summarize the results in a consensus plot, by running the consGSAplot function. For each comparison that we performed, this function creates a plot showing the gene sets that show most significant changes, as identified from consensus GSA. For the highest ranking gene sets, a bar chart is plotted showing how many genes of that gene set are up- or down regulated. To signify which genes have significantly changed expression a p-value cutoff is used. The data used to generate the bar plots are saved in a table and available for further inspection. The table is saved in `comparison name --> GSA level --> df`

```{r, warning=FALSE, results='hide', message=FALSE, fig.height=15, fig.width=10}
Pcutoff <- 0.01 #set p value cutoff
consGSA_plot_data <- list() #create list in which to save the graphs

#generate the graphs and save them in the list, along with the data to generate them
for (i in 1:length(GSA_results)) { #loop through all the GSAs we have
  consGSA_plot_data[[i]] <- list( #the results will be saved in a list
    BP = consGSAplot(resList = GSA_results[[i]][[1]], #create the consensous plot for BP
                     rankScore = 25, 
                     Pcutoff = Pcutoff, 
                     distinct = 'distinct', 
                     title = paste0("Biological Process - rankScore 25", " ,pval < ", Pcutoff)),
    
    MF = consGSAplot(resList =  GSA_results[[i]][[2]], #create the consensous plot for MF
                     rankScore = 25, 
                     Pcutoff = Pcutoff, 
                     distinct = 'distinct', 
                     title = paste0("Molecular Function - rankScore 25" , " ,pval < ", Pcutoff)),
    CC = consGSAplot(resList = GSA_results[[i]][[3]], #create the consensous plot for CC
                     rankScore = 25, 
                     Pcutoff = Pcutoff, 
                     distinct = 'distinct', 
                     title = paste0("Cellular Component - rankScore 25", " ,pval < ", Pcutoff))
  )
  
  names(consGSA_plot_data)[i] <- names(GSA_results)[i] #assign the names to the list of graphs
  
    #For each comparison arrange the three (BP, MF,CC) graphs and print the final plot
    ggsave(ggarrange(consGSA_plot_data[[i]][["BP"]][["plot"]], 
                        consGSA_plot_data[[i]][["MF"]][["plot"]], 
                        consGSA_plot_data[[i]][["CC"]][["plot"]], 
                        ncol=1, 
                        nrow=3, 
                        common.legend = TRUE, 
                        legend = "bottom") %>%
      annotate_figure(top = text_grob((names(comp_list)[i]), 
                                       color = "red", face = "bold", size = 18)),
      path = "../results/Gene_Set_Analysis/Plots" , 
           file=paste0("plot_", names(comp_list[i]), ".pdf"), width = 10, height = 16
      )
}

```

Export the results both in RData file and csv files

```{r, warning = FALSE}
if (!dir.exists("../results/Gene_Set_Analysis/OutputGSA")){dir.create("../results/Gene_Set_Analysis/OutputGSA")}
setwd("../results/Gene_Set_Analysis/OutputGSA")

save(GSA_results, file = "GSA_Result.RData")
save(consGSA_plot_data, file = "consGSA_plot_data.RData")

for (i in 1:length(consGSA_plot_data)){
  for (j in 1:length(consGSA_plot_data[[i]])){
  filename = paste0(names(consGSA_plot_data[[i]])[[j]], "_", names(consGSA_plot_data)[i], ".csv")
  tmp_data = (consGSA_plot_data[[i]][[j]][[1]])
  write.csv2(tmp_data, filename)
}}

```
