==> Urea Utilisation/code/05.Extract_DEG_for_NetworkPlot.Rmd <==
---
title: "Extract DEG for Network plot"
author: "Simone Zaghen"
date: "`r format(Sys.time(), '%d %B, %Y, %H:%M')`"
output: 
  html_document: 
    toc: true
    number_section: true
    toc_float: 
      collapsed: false
      smooth_scroll: true
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../results/Reports") })
---

This script generates eulerr plots and extracts the genes that are differentially expressed among comparisons for the network plot. 

# Load packages and functions

```{r, message=FALSE}
library(tidyverse)
library(eulerr)
library(gridExtra)

source("../code/Functions/mining.R")

select = dplyr::select

```

# Import files

## P value and fold changes

```{r}
load("../results/Comparison_Output/Comparison_Output.RData")

```

## Set the groups 

```{r}

N_Source_116 <- c("OKYL029_U_116_0.1vsOKYL029_AS_116_0.1",
                 "OKYL049_U_116_0.1vsOKYL049_AS_116_0.1",
                 "JFYL007_U_116_0.1vsJFYL007_AS_116_0.1")

N_Source_3 <- c("OKYL029_U_3_0.1vsOKYL029_AS_3_0.1",
                 "OKYL049_U_3_0.1vsOKYL049_AS_3_0.1",
                 "JFYL007_U_3_0.1vsJFYL007_AS_3_0.1")

```

# AS vs U on C/N 116

```{r, fig.width=8}
NS_116 <- mining(df.list = comp_list,
             Group = N_Source_116,
             pval_cutoff = 0.05,
             neg_FC = -1,
             pos_FC = 1)

grid.arrange(NS_116$plot_up, NS_116$plot_dn, ncol = 2)

```

# AS vs U on C/N 3

```{r, fig.width=8}
NS_3 <- mining(df.list = comp_list,
             Group = N_Source_3,
             pval_cutoff = 0.05,
             neg_FC = -1,
             pos_FC = 1)

grid.arrange(NS_3$plot_up, NS_3$plot_dn, ncol = 2)

```

# Save plots

```{r, results='hide'}
if (!dir.exists("../results/Euler_Plot")){dir.create("../results/Euler_Plot")}

png("../results/Euler_Plot/CN_116.png", units = "in", width = 8, height = 4, res = 300)
grid.arrange(NS_116$plot_up, NS_116$plot_dn, ncol = 2)
dev.off()

png("../results/Euler_Plot/CN_3.png", units = "in", width = 8, height = 4, res = 300)
grid.arrange(NS_3$plot_up, NS_3$plot_dn, ncol = 2)
dev.off()

```

# Generate df with info for network plot

```{r}
#Set the up-regulated list
up <- list(OKYL029_116 = NS_116$up$OKYL029_U_116_0.1vsOKYL029_AS_116_0.1$genes,
           OKYL049_116 = NS_116$up$OKYL049_U_116_0.1vsOKYL049_AS_116_0.1$genes,
           JFYL007_116 = NS_116$up$JFYL007_U_116_0.1vsJFYL007_AS_116_0.1$genes,
           OKYL029_3 = NS_3$up$OKYL029_U_3_0.1vsOKYL029_AS_3_0.1$genes,
           OKYL049_3 = NS_3$up$OKYL049_U_3_0.1vsOKYL049_AS_3_0.1$genes,
           JFYL007_3 = NS_3$up$JFYL007_U_3_0.1vsJFYL007_AS_3_0.1$genes)

#Set the downregulated list
down <- list(OKYL029_116 = NS_116$dn$OKYL029_U_116_0.1vsOKYL029_AS_116_0.1$genes,
             OKYL049_116 = NS_116$dn$OKYL049_U_116_0.1vsOKYL049_AS_116_0.1$genes,
             JFYL007_116 = NS_116$dn$JFYL007_U_116_0.1vsJFYL007_AS_116_0.1$genes,
             OKYL029_3 = NS_3$dn$OKYL029_U_3_0.1vsOKYL029_AS_3_0.1$genes,
             OKYL049_3 = NS_3$dn$OKYL049_U_3_0.1vsOKYL049_AS_3_0.1$genes,
             JFYL007_3 = NS_3$dn$JFYL007_U_3_0.1vsJFYL007_AS_3_0.1$genes)

#code to export the data for the netwrok plot
# (https://divenn.tch.harvard.edu)

df_up <- data.frame(Comparison = rep(names(up), sapply(up, length)),
                 Gene = unlist(up), Direction = 1)

df_dn <- data.frame(Comparison = rep(names(down), sapply(down, length)),
                    Gene = unlist(down), Direction = 2)

data_network_plot <- rbind(df_up, df_dn) #create df in which we have data for network plot
data_network_plot <- data_network_plot[order(data_network_plot$Comparison),] #create df in which we have data for network plot

```

# Generate annotation table

This snippet of code will take the intersections between different samples, it will extract the genes in the intersection and will merge them with the annotation in uniprot. 

```{r}
uniprot_table <- read.delim("../data/Uniprot_Yali.txt") %>% #import uniprot table for yali genes
  select(c("Gene.names", "Protein.names"))
uniprot_table$Gene.names <- str_extract(uniprot_table$Gene.names, ".{0,0}YALI1_.{0,7}") #only keep YALI1_ gene names

table_gene_function <- data_network_plot %>% 
  pivot_wider(names_from = Comparison,values_from = Direction) %>% #create a df to check intersections
  merge(uniprot_table, by.x = "Gene", by.y = "Gene.names") #merge with uniprot information 

```

# Export output

```{r}
#save the results and export the dfs in csv files
save(NS_116, file = "../results/Euler_Plot/NS_116.RData")
save(NS_3, file = "../results/Euler_Plot/NS_3.RData")

#export the network plot data
if (!dir.exists("../results/NetworkPlot")) {dir.create("../results/NetworkPlot")}
write.csv2(data_network_plot, file = "../results/NetworkPlot/data_network_plot.csv", row.names = F)
write.csv2(table_gene_function, file = "../results/NetworkPlot/table_gene_function.csv", row.names = F)

```
